Teste_Moonshot/requirements.txt
requests
pandas
numpy
pyyaml
pillow
